+++ ftapi/cursus_stats/get_user_logged_hours.py
@@ -9,7 +9,7 @@ from FtApi import FtApi
 from FtDateTime import is_valid_date, dt_convert
 from FtCursus import DURATION as CURSUS_DURATION, get_cursus_users
 from FtInput import read_input_cursus, read_input_date
-from FtUtils import ft_write_error, ft_write_success
+from utils.io.ft_write_stderr import ft_write_error, ft_write_success
 
 
 def make_attendance_dict(begin_date: str = None, duration: int = 0) -> dict:
