+++ ftapi/users/get_user_quests.py
@@ -3,7 +3,7 @@
 
 from time import sleep
 from FtApi import FtApi
-from FtUtils import ft_write_info
+from utils.io.ft_write_stderr import ft_write_info
 from get_user_id_by_login import get_user_by_login
 
 
@@ -15,6 +15,7 @@ def get_user_quests_by_id(ft_api: FtApi = None, user_id: str = None) -> list:
     if ft_api is None or not isinstance(ft_api, FtApi):
         ft_api = FtApi()
     get_url = f"{ft_api.site}/v2/users/{user_id}/quests"
+    ft_write_info(f"GET {get_url}")
     return ft_api.get(get_url)
 
 
@@ -25,8 +26,9 @@ def get_user_quests_by_login(ft_api: FtApi = None, login: str = None) -> list:
     assert len(login) > 0, "Invalid login: cannot be empty."
     if ft_api is None or not isinstance(ft_api, FtApi):
         ft_api = FtApi()
-    user_id=get_user_by_login(ft_api, login)
-    assert len(user_id) == 1, "Error: Unable to identify unique ID for {login}."
-    user_id=f"{user_id[0]['id']}"
+    user_id = get_user_by_login(ft_api, login)
+    assert len(user_id) == 1, \
+        f"Error: Unable to identify unique ID for {login}."
+    user_id = f"{user_id[0]['id']}"
     sleep(0.8)
-    return get_user_quests_by_id(ft_api=None, user_id=user_id)
+    return get_user_quests_by_id(ft_api=ft_api, user_id=user_id)
