+++ ftapi/utils/io/ft_write_stderr.py
